Fix warning.
(bzr r0.304.87)

+++ src/ring.cpp
@@ -624,7 +624,7 @@ RingScreen::switchToWindow (bool	   toNext)
 	cur++;
     }
 
-    if (cur == windows.size ())
+    if (cur == (int)windows.size ())
 	return;
 
     if (toNext)
